Rust: Migrate pyo3 items out of sqlfluffrs_types
This allows us to compile both `dialects` and `types` once per platform.
The dialects crate took about 5 minutes too compile per Python version.
This way we only compile the dialects once instead of 5 times.
