+++ sqlfluffrs/Cargo.toml
@@ -17,7 +17,11 @@ members = [
     "sqlfluffrs_lexer",
     "sqlfluffrs_parser",
     "sqlfluffrs_types",
+    "sqlfluffrs_python",
 ]
+
+[workspace.package]
+version = "4.0.0"
 
 [features]
 unicode = []
@@ -27,7 +31,7 @@ python = [
     "pyo3",
     "sqlfluffrs_lexer/python",
     "sqlfluffrs_parser/python",
-    "sqlfluffrs_types/python",
+    "sqlfluffrs_python",
 ]
 
 [dependencies]
@@ -56,6 +60,7 @@ sqlfluffrs_types = { path = "sqlfluffrs_types" }
 sqlfluffrs_lexer = { path = "sqlfluffrs_lexer" }
 sqlfluffrs_dialects = { path = "sqlfluffrs_dialects" }
 sqlfluffrs_parser = { path = "sqlfluffrs_parser" }
+sqlfluffrs_python = { path = "sqlfluffrs_python", optional = true }
 
 [dev-dependencies]
 env_logger = "0.11.6"
