+++ sqlfluffrs/sqlfluffrs_lexer/Cargo.toml
@@ -1,5 +1,6 @@
 [package]
 name = "sqlfluffrs_lexer"
+version.workspace = true
 edition = "2021"
 
 [dependencies]
@@ -13,9 +14,10 @@ serde = { version = "1.0.225", features = ["derive"] }
 uuid = { version = "1.18.1", features = ["v4"] }
 sqlfluffrs_types = { path = "../sqlfluffrs_types" }
 sqlfluffrs_dialects = { path = "../sqlfluffrs_dialects" }
+sqlfluffrs_python = { path = "../sqlfluffrs_python", optional = true }
 pyo3 = { version = "0.26.0", optional = true, features = ["hashbrown", "extension-module", "uuid"] }
 pyo3-log = { version = "0.13.0", optional = true }
 
 [features]
 unicode = []
-python = ["unicode", "pyo3", "sqlfluffrs_types/python"]
+python = ["unicode", "pyo3", "sqlfluffrs_python"]
