+++ sqlfluffrs/sqlfluffrs_parser/Cargo.toml
@@ -1,5 +1,6 @@
 [package]
 name = "sqlfluffrs_parser"
+version.workspace = true
 edition = "2021"
 
 [dependencies]
@@ -10,11 +11,12 @@ smallvec = "1.13.2"
 sqlfluffrs_types = { path = "../sqlfluffrs_types" }
 sqlfluffrs_dialects = { path = "../sqlfluffrs_dialects" }
 sqlfluffrs_lexer = { path = "../sqlfluffrs_lexer" }
+sqlfluffrs_python = { path = "../sqlfluffrs_python", optional = true }
 uuid = { version = "1.18.1", features = ["v4"] }
 serde_yaml_ng = "0.10.0"
 regex = { version = "1.11.2", features = ["perf"] }
 pyo3 = { version = "0.26.0", optional = true }
 
 [features]
-python = ["pyo3", "sqlfluffrs_types/python", "sqlfluffrs_lexer/python"]
+python = ["pyo3", "sqlfluffrs_python", "sqlfluffrs_lexer/python"]
 verbose-debug = []
