+++ sqlfluffrs/sqlfluffrs_python/src/lib.rs
@@ -1,0 +1,11 @@
+//! sqlfluffrs_python: scaffold for Python-facing bindings
+#![allow(dead_code)]
+
+pub mod config;
+pub mod marker;
+pub mod slice;
+pub mod templater;
+pub mod token;
+
+// Re-export commonly used types for easier migration.
+pub use token::*;
