+++ sqlfluffrs/sqlfluffrs_types/Cargo.toml
@@ -1,5 +1,6 @@
 [package]
 name = "sqlfluffrs_types"
+version.workspace = true
 edition = "2021"
 
 [dependencies]
@@ -9,11 +10,8 @@ log = "0.4.28"
 serde = { version = "1.0.225", features = ["derive"] }
 uuid = { version = "1.18.1", features = ["v4"] }
 fancy-regex = "0.16.2"
-pyo3 = { version = "0.26.0", optional = true, features = ["hashbrown", "extension-module", "uuid"] }
-pyo3-log = { version = "0.13.0", optional = true }
 bincode = "1.3.3"
 once_cell = "1.21.3"
 
 [features]
 unicode = []
-python = ["unicode", "pyo3"]
