+++ sqlfluffrs/sqlfluffrs_types/src/config/fluffconfig.rs
@@ -12,32 +12,3 @@ impl FluffConfig {
         }
     }
 }
-
-#[cfg(feature = "python")]
-pub mod python {
-    use pyo3::{
-        prelude::*,
-        types::{PyDict, PyDictMethods},
-    };
-
-    use super::FluffConfig;
-
-    #[derive(Clone)]
-    pub struct PyFluffConfig(pub FluffConfig);
-
-    impl<'py> FromPyObject<'py> for PyFluffConfig {
-        fn extract_bound(ob: &Bound<'py, PyAny>) -> PyResult<Self> {
-            let configs = ob.getattr("_configs")?;
-            let configs_dict = configs.downcast::<PyDict>()?;
-            let core = configs_dict.get_item("core").ok().flatten().unwrap();
-            let core_dict = core.downcast::<PyDict>()?;
-            let dialect = core_dict
-                .get_item("dialect")
-                .ok()
-                .flatten()
-                .and_then(|x| x.extract::<String>().ok());
-
-            Ok(Self(FluffConfig::new(dialect, true)))
-        }
-    }
-}
