+++ sqlfluffrs/sqlfluffrs_types/src/lib.rs
@@ -9,28 +9,17 @@ pub mod regex;
 pub mod slice;
 pub mod templater;
 pub mod token;
-
-#[cfg(feature = "python")]
-pub use config::fluffconfig::python::PyFluffConfig;
 pub use config::fluffconfig::FluffConfig;
 pub use grammar_api::{patterns as grammar_patterns, GrammarContext};
 pub use grammar_inst::{GrammarFlags, GrammarId, GrammarInst, GrammarVariant};
 pub use grammar_tables::{
     ChildrenIter, GrammarInstExt, GrammarTables, SimpleHintData, TableMemoryStats, TerminatorsIter,
 };
-#[cfg(feature = "python")]
-pub use marker::python::PyPositionMarker;
 pub use marker::PositionMarker;
 pub use matcher::LexMatcher;
 pub use parser::{ParseMode, RootGrammar, SimpleHint};
 pub use regex::{RegexMode, RegexModeGroup};
 pub use slice::Slice;
-#[cfg(feature = "python")]
-pub use templater::fileslice::python::{PyRawFileSlice, PyTemplatedFileSlice};
 pub use templater::fileslice::{RawFileSlice, TemplatedFileSlice};
-#[cfg(feature = "python")]
-pub use templater::templatefile::python::PyTemplatedFile;
 pub use templater::templatefile::TemplatedFile;
-#[cfg(feature = "python")]
-pub use token::python::{PyCaseFold, PyToken};
 pub use token::{config::TokenConfig, Token};
