+++ sqlfluffrs/sqlfluffrs_types/src/regex.rs
@@ -1,11 +1,8 @@
 use std::fmt::Display;
 
 use fancy_regex::{Regex as FancyRegex, RegexBuilder as FancyRegexBuilder};
-#[cfg(feature = "python")]
-use pyo3::pyclass;
 use regex::{Regex, RegexBuilder};
 
-#[cfg_attr(feature = "python", pyclass)]
 #[derive(Debug, Clone)]
 pub enum RegexModeGroup {
     Index(usize),
